Begin to test.Client find files and record in database, but when server received hashes sent by client and record it in server.db, it seems that some records are redundent.

+++ p2p_class/config.h
@@ -23,7 +23,7 @@ public:
 
 private:
 	std::string m_LocalIP;
-	int m_LocalPort;
+	unsigned short m_LocalPort;
 	std::string m_PeerIP;
 	unsigned short m_PeerPort;
 	unsigned short m_MinPort;
